chore(repo): add prettier infrastructure
this commit adds prettier infrastructure to the project. prettier was
previously a dependency of the project, but was not being used. similar
to other stencil projects, this commit uses ionic's prettier
configuration for consistent formatting. three npm scripts have been
added:
1. `prettier.base` to serve as a single source of truth for common
   settings, such as directories to format
2. `prettier.dry-run` to act as a non-mutating prettier run. used in ci
3. `prettier` to format and write changes back to disk

a ci workflow has been added to kick off a dry-run of prettier. should
any files not be properly formatted, the workflow shall fail